grabbing main
