+++ controllers/api/index.js
@@ -1,12 +1,12 @@
 const router = require('express').Router();
-const customerRoutes = require('./customerRoutes');
-const menuRoutes = require('./menuRoutes');
+// const customerRoutes = require('./customerRoutes');
+// const menuRoutes = require('./menuRoutes');
 const orderRoutes = require('./orderRoutes');
-const reservationsRoutes = require('./reservationRoutes');
+// const reservationRoutes = require('./reservationRoutes');
 
-router.use('/customers', customerRoutes);
-router.use('/menu', menuRoutes);
-router.use('/menu', orderRoutes);
-router.use('/menu', reservationsRoutes);
+// router.use('/customer', customerRoutes);
+// router.use('/menu', menuRoutes);
+router.use('/order', orderRoutes);
+// router.use('/reservation', reservationRoutes);
 
 module.exports = router;
